requirements.txt
streamlit>=1.26.0
openai
SPARQLWrapper
python-dotenv
neo4j
numpy
sentence-transformers
faiss-cpu